+++ src/containerapp/azext_containerapp/tests/latest/utils.py
@@ -10,6 +10,7 @@ from azure.cli.core.azclierror import MutuallyExclusiveArgumentError, RequiredAr
 from azure.cli.testsdk import (JMESPathCheck)
 
 
+
 def create_containerapp_env(test_cls, env_name, resource_group, location=None):
     logs_workspace_name = test_cls.create_random_name(prefix='containerapp-env', length=24)
     logs_workspace_id = test_cls.cmd('monitor log-analytics workspace create -g {} -n {} -l {}'.format(resource_group, logs_workspace_name, TEST_LOCATION)).get_output_in_json()["customerId"]
@@ -25,6 +26,7 @@ def create_containerapp_env(test_cls, env_name, resource_group, location=None):
     while containerapp_env["properties"]["provisioningState"].lower() == "waiting":
         time.sleep(5)
         containerapp_env = test_cls.cmd('containerapp env show -g {} -n {}'.format(resource_group, env_name)).get_output_in_json()
+
 
 def create_and_verify_containerapp_up(
             test_cls,
@@ -74,6 +76,33 @@ def create_and_verify_containerapp_up(
             up_cmd += f" -l {location.upper()}"
             test_cls.cmd(up_cmd)
 
+
+def create_extension_and_custom_location(test_cls, resource_group, connected_cluster_name, custom_location_name):
+    try:
+        connected_cluster = test_cls.cmd(f'az connectedk8s show --resource-group {resource_group} --name {connected_cluster_name}').get_output_in_json()
+        while connected_cluster["connectivityStatus"] == "Connecting":
+            time.sleep(5)
+            connected_cluster = test_cls.cmd(f'az connectedk8s show --resource-group {resource_group} --name {connected_cluster_name}').get_output_in_json()
+
+        connected_cluster_id = connected_cluster.get('id')
+        extension = test_cls.cmd(f'az k8s-extension create'
+                                 f' --resource-group {resource_group}'
+                                 f' --name containerapp-ext'
+                                 f' --cluster-type connectedClusters'
+                                 f' --cluster-name {connected_cluster_name}'
+                                 f' --extension-type "Microsoft.App.Environment" '
+                                 f' --release-train stable'
+                                 f' --auto-upgrade-minor-version true'
+                                 f' --scope cluster'
+                                 f' --release-namespace appplat-ns'
+                                 f' --configuration-settings "Microsoft.CustomLocation.ServiceAccount=default"'
+                                 f' --configuration-settings "appsNamespace=appplat-ns"'
+                                 f' --configuration-settings "clusterName={connected_cluster_name}"'
+                                 f' --configuration-settings "envoy.annotations.service.beta.kubernetes.io/azure-load-balancer-resource-group={resource_group}"').get_output_in_json()
+        test_cls.cmd(f'az customlocation create -g {resource_group} -n {custom_location_name} -l {TEST_LOCATION} --host-resource-id {connected_cluster_id} --namespace appplat-ns -c {extension["id"]}')
+    except:
+        pass
+      
 def verify_containerapp_create_exception(
             test_cls,
             resource_group,